更新dto 符合数据库表内的属性 优化首页加载逻辑 去除空间卡片“默认”按钮 修改为记住用户最后一次点击为默认空间 简化交互逻辑 新增mqtt功能、使用EventBus 代替service+broadcast方案

+++ app/src/main/java/administrator/adapters/AreaCardAdapter.java
@@ -96,7 +96,7 @@ public class AreaCardAdapter extends RecyclerView.Adapter{
                 new LinearLayoutManager(context,LinearLayoutManager.HORIZONTAL,false));
     }
 
-    private void goNextActivity(int areaId) {
+    private void goNextActivity(long areaId) {
         Intent intent = new Intent(context, AreaDetailActivity.class);
         intent.putExtra("area_id",areaId);
         context.startActivity(intent);
